결제 고정값 -> 받아오는 것으로 변경, Data class 수정(Int->Long)

+++ app/src/main/java/com/example/gradfront/data/BookingRequest.kt
@@ -2,6 +2,6 @@ package com.example.gradfront.data
 
 data class BookingRequest(
     val userId: Long,
-    val liveId: Int,
+    val liveId: Long,
     val numberOfTickets: Int
 )
